+++ OpenKh.Command.AnbMaker/Commands/Interfaces/IFbxSourceItemSelector.cs
@@ -1,0 +1,15 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OpenKh.Command.AnbMaker.Commands.Interfaces
+{
+    internal interface IFbxSourceItemSelector
+    {
+        string RootName { get; }
+        string MeshName { get; }
+        string AnimationName { get; }
+    }
+}
